+++ docs/developer-guide/contributing/changes.md
@@ -1,6 +1,8 @@
 ---
 title: Change Requests
-layout: doc_en
+layout: doc
+edit_link: https://github.com/eslint/eslint/edit/master/docs/developer-guide/contributing/changes.md
+
 ---
 <!-- Note: No pull requests accepted for this file. See README.md in the root directory for details. -->
 
@@ -8,16 +10,30 @@ layout: doc_en
 
 If you'd like to request a change to ESLint, please [create a new issue](https://github.com/eslint/eslint/issues/new?template=CHANGE.md) on GitHub. Be sure to include the following information:
 
+如果你想向 ESLint 提出一个变更请求，请在 GitHub [新建一个议题（issue）](https://github.com/eslint/eslint/issues/new?template=CHANGE.md)。一定要包含以下信息：
+
 1. The version of ESLint you are using.
+1. 你正在使用的 ESLint 的版本。
 1. The problem you want to solve.
+1. 你想要解决的问题。
 1. Your take on the correct solution to problem.
+1. 你认为解决问题的正确方法。
 
 If you're requesting a change to a rule, it's helpful to include this information as well:
 
+如果你对一个规则提出变更请求，包含以下信息同样会有所帮助：
+
 1. What you did.
+1. 你做了什么。
 1. What you would like to happen.
+1. 你认为会发生什么。
 1. What actually happened.
+1. 实际上发生了什么。
 
 Please include as much detail as possible to help us properly address your issue. If we need to triage issues and constantly ask people for more detail, that's time taken away from actually fixing issues. Help us be as efficient as possible by including a lot of detail in your issues.
 
+请包括尽可能的多的细节来帮助我们正确地处理你的议题（issue)。如果我们需要对议题（issue)进行分类而且时常会询问更多的细节，那么实际修复这些问题将需要更多的时间。为了使我们尽可能的高效，请在你的议题（issue）中包含更多的细节。
+
 **Note:** If you just have a question that won't necessarily result in a change to ESLint, such as asking how something works or how to contribute, please use the [mailing list](https://groups.google.com/group/eslint) or [chat](https://gitter.im/eslint/eslint) instead of filing an issue.
+
+**注意：**如果你只是有一个问题，且该问题并不一定会导致 ESLint 发生改变，比如询问一些工作原理或者如何做出贡献，那么请使用 [邮件列表（mailing list）](https://groups.google.com/group/eslint) 或[聊天室（chat）](https://gitter.im/eslint/eslint) 而不是提交一个议题（issue）。